Fix incorrect array indexing in index_array
Solves access violations when compiling with MSVC++ 14.0.

+++ src/operator/contrib/index_array-inl.h
@@ -40,10 +40,10 @@ struct IndexArrayKernel {
                                   int64_t* out_data,
                                   const int n,
                                   const int64_t* workspace) {
-    for (int j = 0; j < n; j++) {
-      int64_t upper = workspace[2 * j];
-      int64_t lower = workspace[2 * j + 1];
-      KERNEL_ASSIGN(out_data[i * n + j], req, (i % upper) / lower);
+    for (ptrdiff_t j = 0; j < n; j++) {
+      int64_t upper = workspace[ptrdiff_t(2) * j];
+      int64_t lower = workspace[ptrdiff_t(2) * j + ptrdiff_t(1)];
+      KERNEL_ASSIGN(out_data[ptrdiff_t(i) * ptrdiff_t(n) + j], req, (i % upper) / lower);
     }
   }
 };
@@ -55,8 +55,8 @@ struct IndexArrayDefaultKernel {
                                   const int ndim,
                                   const dim_t* shape) {
     int64_t index = i;
-    for (int j = ndim - 1; j >= 0; j--) {
-      KERNEL_ASSIGN(out_data[i * ndim + j], req, index % shape[j]);
+    for (ptrdiff_t j = ndim - 1; j >= 0; j--) {
+      KERNEL_ASSIGN(out_data[ptrdiff_t(i) * ptrdiff_t(ndim) + j], req, index % shape[j]);
       index /= shape[j];
     }
   }
@@ -84,8 +84,8 @@ inline void IndexArrayBuildSelectedAxesWorkspace(const mxnet::Tuple<int> &axes,
     // Make sure that the axis is between 0 and ndim.
     const int axis = ((axes[i] % ndim) + ndim) % ndim;
 
-    workspace[2 * i] = index_products[axis];
-    workspace[2 * i + 1] = index_products[axis + 1];
+    workspace[ptrdiff_t(2) * ptrdiff_t(i)] = index_products[axis];
+    workspace[ptrdiff_t(2) * ptrdiff_t(i) + ptrdiff_t(1)] = index_products[axis + 1];
   }
 }
